Fix UICR size to accomodate gdb.
We don't know the actual UICR size.  The datasheet documents registers up to 0x210.  So 0x100 is clearly incorrect.

gdb has trouble issuing erase commands if the size isn't aligned to the base address.  This is fixed in upstream gdb, but not released yet.

So the prudent thing is select the next aligned size above 0x210 and use that.  So 0x400.